+++ packages/sdk38/types/index.d.ts
@@ -4,12 +4,17 @@ export { pubkeyToAddress, rawSecp256k1PubkeyToAddress } from "./address";
 export { Coin, coin, coins } from "./coins";
 export {
   Account,
+  assertIsPostTxSuccess,
   Block,
   BlockHeader,
   CosmosClient,
   GetSequenceResult,
   IndexedTx,
+  isPostTxFailure,
+  isPostTxSuccess,
+  PostTxFailure,
   PostTxResult,
+  PostTxSuccess,
   SearchByHeightQuery,
   SearchByIdQuery,
   SearchBySentFromOrToQuery,